requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes -o requirements.txt
absl-py==2.2.1
    # via
    #   keras
    #   tensorboard
    #   tensorflow
asttokens==3.0.0
    # via stack-data
astunparse==1.6.3
    # via tensorflow
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
colorama==0.4.6 ; sys_platform == 'win32'
    # via ipython
contourpy==1.3.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
decorator==5.2.1
    # via ipython
executing==2.2.0
    # via stack-data
filelock==3.18.0
    # via torch
flatbuffers==25.2.10
    # via tensorflow
fonttools==4.56.0
    # via matplotlib
fsspec==2025.3.2
    # via torch
gast==0.6.0
    # via tensorflow
google-pasta==0.2.0
    # via tensorflow
grpcio==1.71.0
    # via
    #   tensorboard
    #   tensorflow
h5py==3.13.0
    # via
    #   keras
    #   tensorflow
idna==3.10
    # via requests
imageio==2.37.0
    # via gan-osuprproject
ipython==9.0.2
    # via gan-osuprproject
ipython-pygments-lexers==1.1.1
    # via ipython
jedi==0.19.2
    # via ipython
jinja2==3.1.6
    # via torch
keras==3.9.1
    # via tensorflow
kiwisolver==1.4.8
    # via matplotlib
libclang==18.1.1
    # via tensorflow
markdown==3.7
    # via tensorboard
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
matplotlib==3.10.1
    # via gan-osuprproject
matplotlib-inline==0.1.7
    # via ipython
mdurl==0.1.2
    # via markdown-it-py
ml-dtypes==0.5.1
    # via
    #   keras
    #   tensorflow
mpmath==1.3.0
    # via sympy
namex==0.0.8
    # via keras
networkx==3.4.2
    # via torch
numpy==2.1.3
    # via
    #   contourpy
    #   h5py
    #   imageio
    #   keras
    #   matplotlib
    #   ml-dtypes
    #   pandas
    #   tensorboard
    #   tensorflow
nvidia-cublas-cu12==12.4.5.8 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.4.127 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cuda-nvrtc-cu12==12.4.127 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cuda-runtime-cu12==12.4.127 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cudnn-cu12==9.1.0.70 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cufft-cu12==11.2.1.3 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-curand-cu12==10.3.5.147 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cusolver-cu12==11.6.1.9 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cusparse-cu12==12.3.1.170 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.6.2 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-nccl-cu12==2.21.5 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-nvjitlink-cu12==12.4.127 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via
    #   nvidia-cufft-cu12
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.4.127 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
opt-einsum==3.4.0
    # via tensorflow
optree==0.14.1
    # via keras
packaging==24.2
    # via
    #   keras
    #   matplotlib
    #   tensorboard
    #   tensorflow
pandas==2.2.3
    # via gan-osuprproject
parso==0.8.4
    # via jedi
pexpect==4.9.0 ; sys_platform != 'emscripten' and sys_platform != 'win32'
    # via ipython
pillow==11.1.0
    # via
    #   imageio
    #   matplotlib
prompt-toolkit==3.0.50
    # via ipython
protobuf==5.29.4
    # via
    #   tensorboard
    #   tensorflow
ptyprocess==0.7.0 ; sys_platform != 'emscripten' and sys_platform != 'win32'
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pygments==2.19.1
    # via
    #   ipython
    #   ipython-pygments-lexers
    #   rich
pyparsing==3.2.3
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytz==2025.2
    # via pandas
requests==2.32.3
    # via tensorflow
rich==13.9.4
    # via keras
setuptools==78.1.0
    # via
    #   tensorboard
    #   tensorflow
    #   torch
six==1.17.0
    # via
    #   astunparse
    #   google-pasta
    #   python-dateutil
    #   tensorboard
    #   tensorflow
stack-data==0.6.3
    # via ipython
sympy==1.13.1
    # via torch
tensorboard==2.19.0
    # via tensorflow
tensorboard-data-server==0.7.2
    # via tensorboard
tensorflow==2.19.0
    # via gan-osuprproject
tensorflow-io-gcs-filesystem==0.37.1 ; python_full_version < '3.12'
    # via tensorflow
termcolor==2.5.0
    # via tensorflow
torch==2.6.0
    # via gan-osuprproject
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
triton==3.2.0 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
typing-extensions==4.13.0
    # via
    #   ipython
    #   optree
    #   tensorflow
    #   torch
tzdata==2025.2
    # via pandas
urllib3==2.3.0
    # via requests
wcwidth==0.2.13
    # via prompt-toolkit
werkzeug==3.1.3
    # via tensorboard
wheel==0.45.1
    # via astunparse
wrapt==1.17.2
    # via tensorflow
